+++ Tests/GuiStracini.SDKBuilder.UnitTests/DummyRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace GuiStracini.SDKBuilder.UnitTests;
+
+using Routing;
+
+[EndpointRoute("something/{Dummy}")]
+public class DummyRequest : BaseRequest
+{
+    public string Dummy { get; set; }
+}
