+++ Tests/GuiStracini.SDKBuilder.UnitTests/GuiStracini.SDKBuilder.UnitTests.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <ProjectGuid>FAE04EC0-301F-11D3-BF4B-00C04F79EFBC</ProjectGuid>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
   </PropertyGroup>
